About last commit of Brain.js, as I said, I had some changes. The first of which, and probably most important one, is that I have made the characters of all characters on the screen into images. seen in the last couple of commits.
I have also made the tile.plant into a variable containing the object of the plants, instead of just a number.
This has always been the intead, but I must have made a mistake somewhere along the way. This should be fixed now.

I have made some changes to how, the top and left, position of the objects is calculated. Explanation:
Since the image itself takes up 50% of a tile, there should only be timed an additional 0.25, on both left and top,
Since this makes up 100%. Also, before there where an addition of 0.5, vh / vw, this has been changed to a full 1.
Since this is clearly the right amount, seeing as this is the padding of the simulation box.